# perf: optimize serialize_path