This is not just a performance improvement, but a bug fix as well. I used to not be possible to get to the root directory on linux, but now that works.

The loop I measured in #344 went from ~880ms down to ~40ms with this. There is still a lot of room for improvement, but this is already huge.

Splitting `working_path` with `last_index_of()` had identical performance to `utils.split_path()`, but using the built-in function is less code.

Creating `filename` and `extension` now only contributes ~1ms to the total time instead of ~20ms.