@dyphire 2022-10-26
There is a bug: the root directory to drive cannot be returned on Windows.