@christoph-heinrich 2022-10-26
I used to not be able to go to the root directory, but it works now.
I don't have a windows installation to test with.
Maybe removing `working_path` and just using `normal_path` will work?
```diff
diff --git a/scripts/uosc/lib/utils.lua b/scripts/uosc/lib/utils.lua
index 821ec8e..b0ae977 100644
--- a/scripts/uosc/lib/utils.lua
+++ b/scripts/uosc/lib/utils.lua
@@ -198,10 +198,7 @@ function serialize_path(path)
 	if not path or is_protocol(path) then return end
 
 	local normal_path = normalize_path(path)
-	-- normalize_path() already strips slashes, but leaves trailing backslash
-	-- for windows drive letters, but we don't need it here.
-	local working_path = normal_path:sub(#normal_path) == '\\' and normal_path:sub(1, #normal_path - 1) or normal_path
-	local dirname, basename = utils.split_path(working_path)
+	local dirname, basename = utils.split_path(normal_path)
 	local dot_i = last_index_of(basename, '.')
 
 	return {

```