@dyphire 2022-10-26
> I used to not be able to go to the root directory, but it works now. I don't have a windows installation to test with. Maybe removing `working_path` and just using `normal_path` will work?

No, the patch does not work.