@christoph-heinrich 2022-10-26
Then I can only suggest adding in some `print` statements to see where things go wrong. Maybe `utils.split_path()` doesn't handle that properly?