@dyphire 2022-10-26
> Then I can only suggest adding in some `print` statements to see where things go wrong. Maybe `utils.split_path()` doesn't handle that properly?

only need to apply the following handle to work normally:
```lua
local dirname, basename = utils.split_path(working_path)
if dirname:match('^[A-Z]+:$') then dirname = nil end
```