@christoph-heinrich 2022-10-26
So it's not possible to switch drives on windows? Didn't know that.
I removed `working_path` and added a backslash to the match instead.