@dyphire 2022-10-26
>  added a backslash to the match instead.

This is wrong. Matching the backslash will cause failure. The match I give can work normally.