@dyphire 2022-10-26
The `working_path` seems necessary. This match does not work properly after you delete it. That's how it works.