@christoph-heinrich 2022-10-26
Did you test the version that had the backslash in the match? Because I don't understand how the backslash removal is necessary but also matching on backslash breaks things because there is no backslash to match...