@dyphire 2022-10-26
> Did you test the version that had the backslash in the match? Because I don't understand how the backslash removal is necessary but also matching on backslash breaks things because there is no backslash to match...

Why do you think it works?

https://user-images.githubusercontent.com/61936050/198110769-35a579e5-9d78-4147-9f8c-bbe346298706.mp4

As I said at the beginning, matching backslashes doesn't work as expected.