@dyphire 2022-10-26
> Oh that's what you mean. I encountered what I think was basically the same failure on linux, so I solved it in the same way. Does that work?

It works.