@christoph-heinrich 2022-10-26
Now I wonder if that will also work, because I noticed that basename is an empty string when serializing root. Would be a bit simpler.
``` diff
diff --git a/scripts/uosc/lib/utils.lua b/scripts/uosc/lib/utils.lua
index 208f73f..c78334f 100644
--- a/scripts/uosc/lib/utils.lua
+++ b/scripts/uosc/lib/utils.lua
@@ -199,8 +199,7 @@ function serialize_path(path)
 
 	local normal_path = normalize_path(path)
 	local dirname, basename = utils.split_path(normal_path)
-	if state.os == 'windows' and normal_path:match('^[A-Z]:\\$') then dirname = nil
-	elseif normal_path == '/' then dirname = nil end
+	if basename == '' then dirname = nil end
 	local dot_i = last_index_of(basename, '.')
 
 	return {
```