@dyphire 2022-10-26
> Now I wonder if that will also work, because I noticed that basename is an empty string when serializing root. Would be a bit simpler.

Ah, it really works. Well done.